Render product summary view (missing images)

+++ App/Features/Foundation/Domain/Domain/Model/Product/Product+Price.swift
@@ -2,10 +2,10 @@ import Foundation
 
 extension Product {
   public struct Price {
-    public let amount: Decimal
+    public let amount: Double
     public let locale: Locale
     
-    public init(amount: Decimal,
+    public init(amount: Double,
                 locale: Locale)
     {
       self.amount = amount
