+++ App/Features/Product/Listing/Listing/Listing/Infrastructure/Extension/Double/Double+String.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-extension Decimal {
+extension Double {
   func toString() -> String {
     return String(describing: self)
   }
